CHORE move the test cases located in the files convert_map_test.go and visit_map_test.go into the file helper_test.go.
